# openapi.json
{"openapi":"3.1.0","info":{"title":"Seixas AI","description":"API de Validação Automatizada de Acordos de Reembolso em Saúde com Base em Norma Interna","version":"1.0.0"},"paths":{"/processes/upload":{"post":{"tags":["Processos Judiciais"],"summary":"Upload Judicial Process","description":"Recebe o arquivo PDF do processo judicial e agenda o processamento assíncrono.","operationId":"upload_judicial_process_processes_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_judicial_process_processes_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProcessSummaryResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/active":{"get":{"tags":["Gestão de Normas Internas"],"summary":"Get Active Policy","description":"Retorna a única versão da norma interna vigente (status = 'ACTIVE') cadastrada.","operationId":"get_active_policy_policies_active_get","parameters":[{"name":"tenant_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tenant Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/draft":{"post":{"tags":["Gestão de Normas Internas"],"summary":"Create Policy Draft","description":"Cria uma nova versão da norma em estado DRAFT a partir de regras estruturadas (JSON).","operationId":"create_policy_draft_policies_draft_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDraftPolicyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/upload-pdf":{"post":{"tags":["Gestão de Normas Internas"],"summary":"Upload Policy Pdf","description":"Recebe o PDF do Manual / Norma Interna submetido pelo usuário, extrai 100% dos critérios\ndinamicamente sem regras prévias em código e salva como DRAFT para revisão humana.","operationId":"upload_policy_pdf_policies_upload_pdf_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_policy_pdf_policies_upload_pdf_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/{version_id}/diff":{"get":{"tags":["Gestão de Normas Internas"],"summary":"Get Policy Diff For Version","description":"Calcula a diferença semântica entre a norma atualmente ACTIVE e a versão informada.","operationId":"get_policy_diff_for_version_policies__version_id__diff_get","parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"string","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/diff":{"get":{"tags":["Gestão de Normas Internas"],"summary":"Get Policy Diff","description":"Calcula a diferença semântica e lógica entre duas versões específicas.","operationId":"get_policy_diff_policies_diff_get","parameters":[{"name":"source_version_id","in":"query","required":true,"schema":{"type":"string","title":"Source Version Id"}},{"name":"target_version_id","in":"query","required":true,"schema":{"type":"string","title":"Target Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/{version_id}/activate":{"post":{"tags":["Gestão de Normas Internas"],"summary":"Activate Policy Version Path","description":"Transição Atômica via path param: Inativa a norma atual e define a nova versão como a ÚNICA 'ACTIVE'.","operationId":"activate_policy_version_path_policies__version_id__activate_post","parameters":[{"name":"version_id","in":"path","required":true,"schema":{"type":"string","title":"Version Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/policies/activate":{"post":{"tags":["Gestão de Normas Internas"],"summary":"Activate Policy Version","description":"Transição Atômica via JSON body.","operationId":"activate_policy_version_policies_activate_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ActivatePolicyRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/hitl/queue":{"get":{"tags":["Human in the Loop"],"summary":"List Hitl Queue","description":"Lista a fila prioritária de processos que exigem revisão humana.","operationId":"list_hitl_queue_hitl_queue_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/HITLReviewItem"},"type":"array","title":"Response List Hitl Queue Hitl Queue Get"}}}}}}},"/hitl/{review_id}/resolve":{"post":{"tags":["Human in the Loop"],"summary":"Resolve Hitl Item","description":"Registra a intervenção e decisão humana definitiva.","operationId":"resolve_hitl_item_hitl__review_id__resolve_post","parameters":[{"name":"review_id","in":"path","required":true,"schema":{"type":"string","title":"Review Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OperatorResolutionPayload"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/health":{"get":{"tags":["Infraestrutura"],"summary":"Health Check","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"ActivatePolicyRequest":{"properties":{"version_id":{"type":"string","title":"Version Id"},"tenant_id":{"type":"string","title":"Tenant Id","default":"tenant_saude_001"}},"type":"object","required":["version_id"],"title":"ActivatePolicyRequest"},"Body_upload_judicial_process_processes_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"cnj_number":{"type":"string","title":"Cnj Number"},"beneficiary_name":{"type":"string","title":"Beneficiary Name"},"operator_name":{"type":"string","title":"Operator Name"}},"type":"object","required":["file","cnj_number","beneficiary_name","operator_name"],"title":"Body_upload_judicial_process_processes_upload_post"},"Body_upload_policy_pdf_policies_upload_pdf_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"},"version":{"type":"string","title":"Version","default":"2026.2"},"policy_name":{"type":"string","title":"Policy Name","default":"Manual de Acordos de Saúde"},"tenant_id":{"type":"string","title":"Tenant Id","default":"tenant_saude_001"}},"type":"object","required":["file"],"title":"Body_upload_policy_pdf_policies_upload_pdf_post"},"CreateDraftPolicyRequest":{"properties":{"policy_name":{"type":"string","title":"Policy Name","default":"Manual de Acordos"},"version":{"type":"string","title":"Version"},"rules":{"items":{"additionalProperties":true,"type":"object"},"type":"array","title":"Rules"},"tenant_id":{"type":"string","title":"Tenant Id","default":"tenant_saude_001"}},"type":"object","required":["version","rules"],"title":"CreateDraftPolicyRequest"},"HITLReviewItem":{"properties":{"review_id":{"type":"string","title":"Review Id"},"process_id":{"type":"string","title":"Process Id"},"cnj_number":{"type":"string","title":"Cnj Number"},"reason":{"type":"string","title":"Reason"},"status":{"type":"string","title":"Status"},"created_at":{"type":"string","title":"Created At"}},"type":"object","required":["review_id","process_id","cnj_number","reason","status","created_at"],"title":"HITLReviewItem"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"OperatorResolutionPayload":{"properties":{"decision":{"type":"string","title":"Decision"},"operator_notes":{"type":"string","title":"Operator Notes"}},"type":"object","required":["decision","operator_notes"],"title":"OperatorResolutionPayload"},"ProcessSummaryResponse":{"properties":{"process_id":{"type":"string","title":"Process Id"},"cnj_number":{"type":"string","title":"Cnj Number"},"total_pages":{"type":"integer","title":"Total Pages"},"status":{"type":"string","title":"Status"},"verdict":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Verdict"},"summary":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Summary"}},"type":"object","required":["process_id","cnj_number","total_pages","status"],"title":"ProcessSummaryResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}